requirements.txt
openai
python-dotenv
pandas
scikit-learn
matplotlib
seaborn
transformers
torch
sentence-transformers
vaderSentiment
nltk
